Use cached image for VM/bootstrap
This adds a check so that when we're running on a VM, we look for
a cached copy of the compressed file on the default route, which
will be the libvirt or baremetal bridge.

Then we can run the *downloader/httpd containers on the host via
dev-scripts to make testing much faster and more efficient in terms
of bandwidth, but in all other cases we download from the internet
as before.